Clean up dead code.
* Delete legacy cmds listnamespaces / syncer-service
* Delete dead code in pkg/client
* Move utility function from pkg/client to pkg/util/policynodeutil
